CloudBenchMaker/tools/requirements.txt
sage_data_client==0.7.1
requests>=2.31.0
Pillow>=10.0.0
pandas>=2.0.0